+++ lib/dashboard.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:carousel_slider/carousel_slider.dart';
 
-
 class CoalitionDashboard extends StatefulWidget {
   const CoalitionDashboard({super.key});
