+++ pubspec.yaml
@@ -19,6 +19,7 @@ dependencies:
   flex_color_scheme: ^7.2.0
   carousel_slider: ^4.2.1
   url_launcher: ^6.1.12
+  go_router: ^10.0.0
 
 dev_dependencies:
   flutter_launcher_icons: "^0.13.1"
